Move Edit component into its own file
